+++ app/src/engineering/java/com/hedvig/app/feature/onboarding/MockChoosePlanViewModel.kt
@@ -13,6 +13,6 @@ class MockChoosePlanViewModel : ChoosePlanViewModel() {
     }
 
     companion object {
-        var mockData = CHOOSE_PLAN_DATA
+        var mockData = CHOOSE_PLAN_DATA.embarkStories
     }
 }
