+++ app/src/engineering/java/com/hedvig/app/feature/onboarding/OnboardingMockActivity.kt
@@ -2,25 +2,32 @@ package com.hedvig.app.feature.onboarding
 
 import com.hedvig.app.MockActivity
 import com.hedvig.app.choosePlanModule
+import com.hedvig.app.feature.marketpicker.Market
+import com.hedvig.app.feature.marketpicker.MarketProvider
 import com.hedvig.app.feature.onbarding.ChoosePlanViewModel
 import com.hedvig.app.feature.onbarding.MoreOptionsViewModel
 import com.hedvig.app.feature.onbarding.ui.ChoosePlanActivity
 import com.hedvig.app.feature.onbarding.ui.MoreOptionsActivity
 import com.hedvig.app.genericDevelopmentAdapter
+import com.hedvig.app.marketProviderModule
+import com.hedvig.app.mocks.MockMarketProvider
 import com.hedvig.app.onboardingModule
 import org.koin.android.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
 class OnboardingMockActivity : MockActivity() {
-    override val original = listOf(onboardingModule, choosePlanModule)
+    private val marketProvider = MockMarketProvider()
+    override val original = listOf(onboardingModule, choosePlanModule, marketProviderModule)
     override val mocks = listOf(module {
         viewModel<MoreOptionsViewModel> { MockMoreOptionsViewModel() }
         viewModel<ChoosePlanViewModel> { MockChoosePlanViewModel() }
+        single<MarketProvider> { marketProvider }
     })
 
     override fun adapter() = genericDevelopmentAdapter {
         header("Choose Plan")
-        clickableItem("Choose plan") {
+        clickableItem("Choose plan, Market: Norway") {
+            MockMarketProvider.mockedMarket = Market.NO
             startActivity(ChoosePlanActivity.newInstance(this@OnboardingMockActivity))
         }
         header("More Options")
