+++ app/src/engineering/java/com/hedvig/app/mocks/MockPaymentViewModel.kt
@@ -1,6 +1,5 @@
 package com.hedvig.app.mocks
 
-import com.hedvig.android.owldroid.graphql.RedeemReferralCodeMutation
 import com.hedvig.app.feature.profile.ui.payment.PaymentViewModel
 import com.hedvig.app.testdata.feature.payment.PAYIN_STATUS_DATA_ACTIVE
 import com.hedvig.app.testdata.feature.payment.PAYMENT_DATA_TRUSTLY_CONNECTED
