+++ testdata/src/main/java/com/hedvig/app/testdata/feature/onboarding/ChoosePlanData.kt
@@ -1,6 +1,5 @@
 package com.hedvig.app.testdata.feature.onboarding
 
-import com.hedvig.android.owldroid.graphql.ChoosePlanQuery
 import com.hedvig.app.testdata.feature.onboarding.builders.EmbarkStoriesBuilder
 
 val CHOOSE_PLAN_DATA = EmbarkStoriesBuilder().build()
